requirements.txt

streamlit
opencv-python
scikit-learn
numpy
pillow
